server/pyproject.toml
[tool.poetry]
name = "revou-final-project-repositories"
version = "0.1.0"
description = ""
authors = ["Alvido Faisal <alvidofaisal@gmail.com>"]
readme = "README.md"
include = ["src/**/*"]
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.3"
SQLAlchemy = "^2.0.29"
mysqlclient = "^2.2.4"
flask-login = "^0.6.3"
flask-wtf = "^1.2.1"
werkzeug = "^3.0.2"
flask-sqlalchemy = "^2.5.1"
pytest = "^8.2.0"
flask-mail = "^0.9.1"
flask-jwt = "^0.3.2"
email-validator = "^2.1.1"
faker = "^25.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"